Silence gcc warning.
Fix the build with CFLAGS="-Wall -Werror":
|   CC     FSGetErrorText.o
| cc1: warnings being treated as errors
| FSGetErrorText.c: In function ‘CheckErrorMessage’:
| FSGetErrorText.c:43: error: format ‘%d’ expects type ‘int’, but argument 5 has type ‘long unsigned int’

An int is likely to be sufficient for this buffer's size…

+++ test/FSGetErrorText.c
@@ -40,7 +40,7 @@ CheckErrorMessage(FSServer *srv, int code, const char *codestr)
 
     if (!FSGetErrorText(srv, code, buf, sizeof(buf))) {
 	fprintf(stderr, "FSGetErrorText(srv, %s (%d), buf, %d) failed",
-		codestr, code, sizeof(buf));
+		codestr, code, (int) sizeof(buf));
 	return 0;
     }
